Split workspace upstream client protocols
